added navigation generation

+++ src/core/dorcs_file.rs
@@ -5,7 +5,7 @@ use std::{fs, path::PathBuf};
 #[derive(Debug, Clone)]
 pub struct MetaData {
     pub title: String,
-    pub position: Option<i32>,
+    pub position: Option<isize>,
 }
 
 impl MetaData {
@@ -17,7 +17,7 @@ impl MetaData {
     fn parse_metadata(raw_content: &String) -> MetaData {
         let re = Regex::new(r"(?s)^---\s*(.*?)\s*---").unwrap();
         let mut title = String::new();
-        let mut position: Option<i32> = None;
+        let mut position: Option<isize> = None;
 
         if let Some(cap) = re.captures_iter(raw_content.as_str()).next() {
             let meta_str = &cap[1];
@@ -31,7 +31,7 @@ impl MetaData {
                     // Match the key and update the struct fields accordingly
                     match key {
                         "title" => title = value.to_string(),
-                        "position" => position = value.parse::<i32>().ok(),
+                        "position" => position = value.parse::<isize>().ok(),
                         _ => {} // Ignore unknown keys
                     }
                 }
@@ -91,4 +91,3 @@ impl DorcsFile {
         html.unwrap()
     }
 }
-// TODO: Load the files and properly insert them into the navigation tree.
